also use get<float>, replace python dict() by access through __dict__ to prevent terminal escape chars - hopefully, silence "creating shared library"
git-svn-id: http://root.cern.ch/svn/roottest/trunk@1486 7c735d0d-823a-0410-acfd-cd159556d4bf

+++ cint/functionTemplate/printme.C
@@ -4,7 +4,7 @@
 void printme(const t& o) {
    double d=o.get<double>();
    int i=o.get<int>();
-   int j=o.get<double>();
+   int j=o.get<float>();
 
    std::cout << "t now " << d << " " << i << " " << j << std::endl;
    float v = o.getfloat();
